test: improved insert identity tests
