+++ packages/rs-drive/src/drive/identity/insert/add_new_identity/mod.rs
@@ -121,3 +121,104 @@ impl Drive {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::util::test_helpers::setup::{setup_drive, setup_drive_with_initial_state_structure};
+    use dpp::identity::Identity;
+
+    use dpp::block::block_info::BlockInfo;
+    use dpp::fee::fee_result::FeeResult;
+    use dpp::identity::accessors::IdentityGettersV0;
+
+    use dpp::version::PlatformVersion;
+
+    #[test]
+    fn test_insert_and_fetch_identity_first_version() {
+        let platform_version = PlatformVersion::first();
+        let expected_fee_result = FeeResult {
+            storage_fee: 128871000,
+            processing_fee: 2330320,
+            ..Default::default()
+        };
+
+        test_insert_and_fetch_identity(true, platform_version, expected_fee_result);
+    }
+
+    #[test]
+    fn test_insert_and_fetch_identity_latest_version() {
+        let platform_version = PlatformVersion::latest();
+        let expected_fee_result = FeeResult {
+            storage_fee: 128871000,
+            processing_fee: 2330320,
+            ..Default::default()
+        };
+
+        test_insert_and_fetch_identity(true, platform_version, expected_fee_result);
+    }
+
+    #[test]
+    fn test_insert_identity_estimated_costs_first_version() {
+        let platform_version = PlatformVersion::first();
+        let expected_fee_result = FeeResult {
+            storage_fee: 128871000,
+            processing_fee: 11764980,
+            ..Default::default()
+        };
+
+        test_insert_and_fetch_identity(false, platform_version, expected_fee_result);
+    }
+
+    #[test]
+    fn test_insert_identity_estimated_costs_latest_version() {
+        let platform_version = PlatformVersion::latest();
+        let expected_fee_result = FeeResult {
+            storage_fee: 128871000,
+            processing_fee: 11764980,
+            ..Default::default()
+        };
+
+        test_insert_and_fetch_identity(false, platform_version, expected_fee_result);
+    }
+
+    fn test_insert_and_fetch_identity(
+        apply: bool,
+        platform_version: &PlatformVersion,
+        expected_fee_result: FeeResult,
+    ) {
+        let drive = setup_drive(None);
+
+        let transaction = drive.grove.start_transaction();
+
+        drive
+            .create_initial_state_structure(Some(&transaction), platform_version)
+            .expect("expected to create root tree successfully");
+
+        let identity = Identity::random_identity(5, Some(12345), platform_version)
+            .expect("expected a random identity");
+
+        let fee_result = drive
+            .add_new_identity(
+                identity.clone(),
+                false,
+                &BlockInfo::default(),
+                apply,
+                Some(&transaction),
+                platform_version,
+            )
+            .expect("expected to insert identity");
+
+        if apply {
+            let fetched_identity = drive
+                .fetch_full_identity(
+                    identity.id().to_buffer(),
+                    Some(&transaction),
+                    platform_version,
+                )
+                .expect("should fetch an identity")
+                .expect("should have an identity");
+            assert_eq!(identity, fetched_identity);
+        }
+        assert_eq!(fee_result, expected_fee_result);
+    }
+}
