+++ packages/rs-drive/src/drive/identity/insert/add_new_identity/v0/mod.rs
@@ -270,80 +270,3 @@ impl Drive {
         Ok(batch_operations)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::util::test_helpers::setup::{setup_drive, setup_drive_with_initial_state_structure};
-    use dpp::identity::Identity;
-
-    use dpp::block::block_info::BlockInfo;
-    use dpp::identity::accessors::IdentityGettersV0;
-
-    use dpp::version::PlatformVersion;
-
-    #[test]
-    fn test_insert_and_fetch_identity_v0() {
-        let drive = setup_drive(None);
-        let platform_version = PlatformVersion::first();
-
-        let transaction = drive.grove.start_transaction();
-
-        drive
-            .create_initial_state_structure(Some(&transaction), platform_version)
-            .expect("expected to create root tree successfully");
-
-        let identity = Identity::random_identity(5, Some(12345), platform_version)
-            .expect("expected a random identity");
-
-        drive
-            .add_new_identity_v0(
-                identity.clone(),
-                false,
-                &BlockInfo::default(),
-                true,
-                Some(&transaction),
-                platform_version,
-            )
-            .expect("expected to insert identity");
-
-        let fetched_identity = drive
-            .fetch_full_identity(
-                identity.id().to_buffer(),
-                Some(&transaction),
-                platform_version,
-            )
-            .expect("should fetch an identity")
-            .expect("should have an identity");
-
-        assert_eq!(identity, fetched_identity);
-    }
-
-    #[test]
-    fn test_insert_identity_v0() {
-        let drive = setup_drive_with_initial_state_structure(None);
-
-        let db_transaction = drive.grove.start_transaction();
-
-        let platform_version = PlatformVersion::latest();
-
-        let identity = Identity::random_identity(5, Some(12345), platform_version)
-            .expect("expected a random identity");
-
-        drive
-            .add_new_identity_v0(
-                identity,
-                false,
-                &BlockInfo::default(),
-                true,
-                Some(&db_transaction),
-                platform_version,
-            )
-            .expect("expected to insert identity");
-
-        drive
-            .grove
-            .commit_transaction(db_transaction)
-            .unwrap()
-            .expect("expected to be able to commit a transaction");
-    }
-}
